Move CompressionFlags definitions to sepearate file.
