+++ src/intcomp/IntCompress.cpp
@@ -40,59 +40,7 @@ using namespace interp;
 using namespace utils;
 using namespace wasm;
 
-namespace {
-
-charstring CollectionFlagsName[] = {"None", "TopLevel", "IntPaths", "All"};
-
-}  // end of anonymous namespace
-
 namespace intcomp {
-
-charstring getName(CollectionFlags Flags) {
-  if (Flags < size(CollectionFlagsName))
-    return CollectionFlagsName[Flags];
-  return "UnknownCollectionFlags";
-}
-
-CompressionFlags::CompressionFlags()
-    : CountCutoff(0),
-      WeightCutoff(0),
-      PatternLengthLimit(10),
-      PatternLengthMultiplier(2),
-      MaxAbbreviations(4096),
-      SmallValueMax(std::numeric_limits<uint8_t>::max()),
-      SmallValueCountCutoff(2),
-      AbbrevFormat(IntTypeFormat::Varuint64),
-      MinimizeCodeSize(true),
-      UseHuffmanEncoding(false),
-      TrimOverriddenPatterns(false),
-      ReassignAbbreviations(true),
-      DefaultFormat(IntTypeFormat::Varint64),
-      LoopSizeFormat(IntTypeFormat::Varuint64),
-      TraceHuffmanAssignments(false),
-      TraceReadingInput(false),
-      TraceReadingIntStream(false),
-      TraceWritingCodeOutput(false),
-      TraceWritingDataOutput(false),
-      TraceCompression(false),
-      TraceIntStreamGeneration(false),
-      TraceCodeGenerationForReading(false),
-      TraceCodeGenerationForWriting(false),
-      TraceInputIntStream(false),
-      TraceIntCounts(false),
-      TraceIntCountsCollection(false),
-      TraceSequenceCounts(false),
-      TraceSequenceCountsCollection(false),
-      TraceInitialAbbreviationAssignments(false),
-      TraceAbbreviationAssignments(false),
-      TraceAbbreviationAssignmentsCollection(false),
-      TraceAssigningAbbreviations(false),
-      TraceCompressedIntOutput(false),
-      TraceAbbrevSelectionSelect(false),
-      TraceAbbrevSelectionCreate(false),
-      TraceAbbrevSelectionDetail(false),
-      TraceAbbrevSelectionProgress(0) {
-}
 
 IntCompressor::IntCompressor(std::shared_ptr<decode::Queue> Input,
                              std::shared_ptr<decode::Queue> Output,
